requirements.txt
crewai
crewai_tools
python-dotenv
langchain-huggingface
langchain
langchain_openai
